move clearFilters function to the rules table context
